requirements.txt
gensim==4.3.2
joblib==1.2.0
matplotlib==3.7.1
nltk==3.8.1
pandas==2.0.2
python_Levenshtein==0.21.1
scikit_learn==1.2.2
torch==2.0.1
